[clang][bytecode] Fix a build failure on aarch64
This broke e.g.
https://lab.llvm.org/buildbot/#/builders/190/builds/11216

+++ clang/lib/AST/ByteCode/InterpBuiltin.cpp
@@ -1950,7 +1950,8 @@ static bool interp__builtin_memcmp(InterpState &S, CodePtr OpPC,
 
   size_t MinBufferSize = std::min(BufferA.byteSize().getQuantity(),
                                   BufferB.byteSize().getQuantity());
-  size_t CmpSize = std::min(MinBufferSize, Size.getZExtValue());
+  size_t CmpSize =
+      std::min(MinBufferSize, static_cast<size_t>(Size.getZExtValue()));
   int Result = std::memcmp(BufferA.Data.get(), BufferB.Data.get(), CmpSize);
   if (Result == 0)
     pushInteger(S, 0, Call->getType());
